fix(frontend): unify category create and edit editors
